refactor: calendar-carousel → calendarCarousel로 폴더 네이밍 변경
