+++ apps/web/src/features/calendarCarousel/model/useWeeklyCarousel.ts
